clean Widget 定義 protocol をファイルで分割しました.
