+++ Sources/GameWidget/Commons/Widget/Widget+PlugIns.swift
@@ -1,0 +1,12 @@
+//
+//  Widget+PlugIns.swift
+//  
+//
+//  Created by rrbox on 2023/02/16.
+//
+
+public extension NodeGenerator where Self: ContextPresentPlugIn, Context: ParameterLessGeneratable {
+    func node() -> Node {
+        self.node(context: Context())
+    }
+}
