+++ Sources/GameWidget/Commons/Widget/Widget.swift
@@ -7,12 +7,6 @@
 
 import SpriteKit
 
-public protocol ContextPresentPlugIn {
-    associatedtype Context: WidgetContextType
-    associatedtype Node: SKNode
-    func node(context: Context) -> Node
-}
-
 public protocol NodeGenerator {
     associatedtype Node: SKNode
     func node() -> Node
@@ -29,11 +23,5 @@ public protocol Widget: NodeGenerator, WidgetListElementType {
 public extension WidgetListElementType where Self: NodeGenerator {
     func addTo(buffer: inout [SKNode]) {
         buffer.append(self.node())
-    }
-}
-
-public extension NodeGenerator where Self: ContextPresentPlugIn, Context: ParameterLessGeneratable {
-    func node() -> Node {
-        self.node(context: Context())
     }
 }
